Revert "Merge pull request #2615 from farodin91/refactor-settings-cluster"
This reverts commit 39d8f7295b5517e35cb75e44b870d7c99de31529, reversing
changes made to f5dc9fdfd8865cf73f0785b5e72bee2b8e038248.
